# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FinEdge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface UserProfile {
  id: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  isPro: boolean;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  createdAt?: string;
}

export interface UserProfileUpdate {
  displayName?: string;
}

export interface QuizPreferences {
  userId: string;
  strategies?: string[];
  /** @nullable */
  capitalRange?: string | null;
  /** @nullable */
  profitTarget?: string | null;
  completed: boolean;
}

export interface QuizPreferencesInput {
  strategies: string[];
  capitalRange: string;
  profitTarget: string;
}

export interface LearningStreak {
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  lastCheckinDate: string | null;
  checkinDates: string[];
}

export interface Lesson {
  id: number;
  title: string;
  explanation: string;
  whyItMatters: string;
  howItWorks: string[];
  tag: string;
  category: string;
  dayIndex: number;
}

export interface LessonPreview {
  id: number;
  title: string;
  tag: string;
  dayIndex: number;
}

export interface StockQuote {
  ticker: string;
  /** @nullable */
  companyName?: string | null;
  price: number;
  change: number;
  changePercent: number;
  previousClose: number;
  previousHigh?: number;
  previousLow?: number;
  /** @nullable */
  volume?: number | null;
  /** @nullable */
  avgVolume?: number | null;
  /** @nullable */
  atr?: number | null;
  isHighVolatility?: boolean;
  sparkline?: number[];
}

export interface WatchlistItem {
  id: number;
  userId: string;
  ticker: string;
  companyName: string;
  addedAt: string;
  quote?: StockQuote;
}

export interface WatchlistInput {
  ticker: string;
  companyName?: string;
}

export interface OHLCCandle {
  time: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface StockCard {
  ticker: string;
  companyName: string;
  price: number;
  changePercent: number;
  previousHigh: number;
  previousLow: number;
  /** @nullable */
  atr?: number | null;
  isHighVolatility?: boolean;
  candles: OHLCCandle[];
}

export interface StockSearchResult {
  ticker: string;
  companyName: string;
  /** @nullable */
  sector?: string | null;
}

export interface InsightCard {
  ticker: string;
  companyName: string;
  summary: string;
  label: string;
  /** @nullable */
  changePercent30d?: number | null;
  generatedAt: string;
}

export interface MarketInsights {
  opportunity: InsightCard;
  slowdown: InsightCard;
}

export interface PaperPortfolio {
  userId: string;
  cashBalance: number;
  totalValue: number;
  unrealizedPnl: number;
  startingBalance: number;
}

export interface PaperPosition {
  id: number;
  ticker: string;
  companyName: string;
  shares: number;
  avgCost: number;
  currentPrice: number;
  totalValue: number;
  unrealizedPnl: number;
  unrealizedPnlPercent: number;
}

export type PaperTradeInputAction = typeof PaperTradeInputAction[keyof typeof PaperTradeInputAction];


export const PaperTradeInputAction = {
  buy: 'buy',
  sell: 'sell',
} as const;

export interface PaperTradeInput {
  ticker: string;
  shares: number;
  action: PaperTradeInputAction;
}

export interface PaperTrade {
  id: number;
  ticker: string;
  /** @nullable */
  companyName?: string | null;
  shares: number;
  price: number;
  action: string;
  total: number;
  executedAt: string;
}

export interface PaperTradeResult {
  success: boolean;
  trade: PaperTrade;
  portfolio: PaperPortfolio;
  /** @nullable */
  message?: string | null;
}

export interface PerformancePoint {
  date: string;
  value: number;
}

export interface PortfolioHolding {
  id: number;
  userId: string;
  ticker: string;
  companyName: string;
  shares: number;
  avgCost: number;
  addedAt: string;
  /** @nullable */
  currentPrice?: number | null;
  /** @nullable */
  totalValue?: number | null;
  /** @nullable */
  gainLoss?: number | null;
  /** @nullable */
  gainLossPercent?: number | null;
  sparkline?: number[];
}

export interface PortfolioHoldingInput {
  ticker: string;
  companyName?: string;
  shares: number;
  avgCost: number;
}

export interface PortfolioHoldingUpdate {
  shares?: number;
  avgCost?: number;
}

export interface AllocationSlice {
  ticker: string;
  companyName: string;
  value: number;
  percentage: number;
}

export interface PortfolioSummary {
  totalValue: number;
  holdings: AllocationSlice[];
}

export interface SubscriptionStatus {
  isPro: boolean;
  /** @nullable */
  status: string | null;
  /** @nullable */
  currentPeriodEnd?: string | null;
  /** @nullable */
  priceId?: string | null;
}

export interface CheckoutInput {
  priceId: string;
}

export interface CheckoutSession {
  url: string;
}

export interface PortalSession {
  url: string;
}

export interface PriceInfo {
  id: string;
  unitAmount: number;
  currency: string;
  /** @nullable */
  interval: string | null;
}

export interface ProductWithPrice {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  prices?: PriceInfo[];
}

export type SearchStocksParams = {
q: string;
};