# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FinEdge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stocks/quote/{ticker}:
    get:
      operationId: getStockQuote
      tags: [stocks]
      summary: Get current stock quote
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockQuote"
...
  /paper-trade/portfolio:
    get:
      operationId: getPaperPortfolio
      tags: [paper-trade]
      summary: Get paper trading portfolio
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Paper portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperPortfolio"
  /paper-trade/positions:
    get:
      operationId: getPaperPositions
      tags: [paper-trade]
      summary: Get open paper trading positions
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Paper positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperPosition"
  /paper-trade/trade:
    post:
      operationId: executePaperTrade
      tags: [paper-trade]
      summary: Execute a paper trade (buy or sell)
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperTradeInput"
      responses:
        "200":
          description: Trade result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTradeResult"
  /paper-trade/history:
    get:
      operationId: getPaperTradeHistory
      tags: [paper-trade]
      summary: Get paper trade history
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperTrade"
  /paper-trade/performance:
    get:
      operationId: getPaperPerformance
      tags: [paper-trade]
      summary: Get portfolio value over time for performance chart
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Performance data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerformancePoint"
  /portfolio/holdings:
    get:
      operationId: getPortfolioHoldings
      tags: [portfolio]
      summary: Get user portfolio holdings
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Portfolio holdings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioHolding"
    post:
      operationId: addPortfolioHolding
      tags: [portfolio]
      summary: Add a holding to portfolio
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioHoldingInput"
      responses:
        "201":
          description: Added holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioHolding"
  /portfolio/holdings/{id}:
    patch:
      operationId: updatePortfolioHolding
      tags: [portfolio]
      summary: Update portfolio holding shares
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioHoldingUpdate"
      responses:
        "200":
          description: Updated holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioHolding"
    delete:
      operationId: removePortfolioHolding
      tags: [portfolio]
      summary: Remove a holding from portfolio
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /portfolio/summary:
    get:
      operationId: getPortfolioSummary
      tags: [portfolio]
      summary: Get portfolio allocation summary for pie chart
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Portfolio summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioSummary"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    StockQuote:
      type: object
      required: [ticker, price, change, changePercent, previousClose]
      properties:
        ticker:
          type: string
        companyName:
          type: ["string", "null"]
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        previousClose:
          type: number
        previousHigh:
          type: number
        previousLow:
          type: number
        volume:
          type: ["integer", "null"]
        avgVolume:
          type: ["integer", "null"]
        atr:
          type: ["number", "null"]
        isHighVolatility:
          type: boolean
        sparkline:
          type: array
          items:
            type: number
...
    PaperPortfolio:
      type: object
      required: [userId, cashBalance, totalValue, unrealizedPnl, startingBalance]
      properties:
        userId:
          type: string
        cashBalance:
          type: number
        totalValue:
          type: number
        unrealizedPnl:
          type: number
        startingBalance:
          type: number
    PaperPosition:
      type: object
      required: [id, ticker, companyName, shares, avgCost, currentPrice, totalValue, unrealizedPnl, unrealizedPnlPercent]
      properties:
        id:
          type: integer
        ticker:
          type: string
        companyName:
          type: string
        shares:
          type: number
        avgCost:
          type: number
        currentPrice:
          type: number
        totalValue:
          type: number
        unrealizedPnl:
          type: number
        unrealizedPnlPercent:
          type: number
    PaperTradeInput:
      type: object
      required: [ticker, shares, action]
      properties:
        ticker:
          type: string
        shares:
          type: number
        action:
          type: string
          enum: [buy, sell]
    PaperTradeResult:
      type: object
      required: [success, trade, portfolio]
      properties:
        success:
          type: boolean
        trade:
          $ref: "#/components/schemas/PaperTrade"
        portfolio:
          $ref: "#/components/schemas/PaperPortfolio"
        message:
          type: ["string", "null"]
    PaperTrade:
      type: object
      required: [id, ticker, shares, price, action, total, executedAt]
      properties:
        id:
          type: integer
        ticker:
          type: string
        companyName:
          type: ["string", "null"]
        shares:
          type: number
        price:
          type: number
        action:
          type: string
        total:
          type: number
        executedAt:
          type: string
    PerformancePoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number
    PortfolioHolding:
      type: object
      required: [id, userId, ticker, companyName, shares, avgCost, addedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        ticker:
          type: string
        companyName:
          type: string
        shares:
          type: number
        avgCost:
          type: number
        addedAt:
          type: string
        currentPrice:
          type: ["number", "null"]
        totalValue:
          type: ["number", "null"]
        gainLoss:
          type: ["number", "null"]
        gainLossPercent:
          type: ["number", "null"]
        sparkline:
          type: array
          items:
            type: number
    PortfolioHoldingInput:
      type: object
      required: [ticker, shares, avgCost]
      properties:
        ticker:
          type: string
        companyName:
          type: string
        shares:
          type: number
        avgCost:
          type: number
    PortfolioHoldingUpdate:
      type: object
      properties:
        shares:
          type: number
        avgCost:
          type: number
    PortfolioSummary:
      type: object
      required: [totalValue, holdings]
      properties:
        totalValue:
          type: number
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/AllocationSlice"
    AllocationSlice:
      type: object
      required: [ticker, companyName, value, percentage]
      properties:
        ticker:
          type: string
        companyName:
          type: string
        value:
          type: number
        percentage:
          type: number
...